% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STVCox.R
\name{sim_data_multi_fun_unstra}
\alias{sim_data_multi_fun_unstra}
\title{Simulate unstratified survival data with time-varying effects}
\usage{
sim_data_multi_fun_unstra(
  N = 1000,
  seed = 123,
  censor.max = 10,
  end.time = 3,
  fun.list = NULL,
  gamma0 = 0.5,
  z.type = "Ind",
  z.para = 0.5
)
}
\arguments{
\item{N}{sample size}

\item{seed}{random seed}

\item{censor.max}{censoring time}

\item{end.time}{the study end time}

\item{fun.list}{a list of time-varying functions for covariates}

\item{gamma0}{the baseline hazard (a constant)}

\item{z.type}{a character indicating correlation structure of covariate variables. 
the possible options include: "Ind" for "independent",
"AR1" for "order 1 auto regressive" and "CS" for "compound symmetry".}

\item{z.para}{a parameter to specify the covariate correlation according to "z.type".}
}
\value{
a list object of four elements
\describe{
\item{delta}{event or consoring indicator}
\item{z}{covariate matrix}
\item{time}{event time}
\item{event.rate}{the event rate}
\item{fun.list}{true functions}
}
}
\description{
Simulate unstratified survival data with time-varying effects
}
\examples{
dat <- sim_data_multi_fun_unstra()
fit <- with(dat,survival::coxph(Surv(time,delta)~z))
print(fit)
}
\author{
Jian Kang <jiankang@umich.edu> Yuan Yang, Yi Li
}
