% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STVCox-package.R
\docType{package}
\name{STVCox-package}
\alias{STVCox}
\alias{STVCox-package}
\title{Soft-thresholding operator for sparse time varying 
effects in Cox model}
\description{
This package implements the parameter estimation and statistical 
inferences on sparse time-varying effects in survival models based on a soft-thresholding approach.
The package has two functions:
\link{STVCox_fit} for model fitting and \link{sim_data_multi_fun_unstra} for simulating data
}
\keyword{internal}
